Re-order and add more Homebrew tools
Git we need right at the start, so that the installation script can
symlink the diff highlight script.

Telnet is just useful and we don't need curl immediately so they can
both go in the secondary tools block.

CMake we need to build the vterm module for Emacs

And WhatsApp on the desktop is useful too.